+++ BuberDinner.Domain/Common/ValueObjects/AverageRating.cs
@@ -4,17 +4,17 @@ namespace BuberDinner.Domain.Common.ValueObjects
 {
     public sealed class AverageRating : ValueObject
     {
-        public float Value { get; private set; }
+        public double Value { get; private set; }
         public int NumRatings { get; private set; }
 
-        private AverageRating(float value, int numRatings)
+        private AverageRating(double value, int numRatings)
         {
             Value = value;
             NumRatings = numRatings;
         }
 
         public static AverageRating CreateNew(
-            float value = 0,
+            double value = 0,
             int numRatings = 0)
         {
             return new AverageRating(value, numRatings);
